Secure interactive outgoing references
