+++ admin/app/pagepresser/InteractiveHtmlCleaner.scala
@@ -15,15 +15,10 @@ object InteractiveHtmlCleaner extends HtmlCleaner with implicits.WSRequests {
 
   override def clean(document: Document) = {
     universalClean(document)
-    //removeScripts(document)
+    removeScripts(document)
     createSimplePageTracking(document)
     removeByTagName(document, "noscript")
-
-    // prep. for HTTPS
-//    val tmpDoc = Jsoup.parse(secureSource(document.html()))
-//    document.head().replaceWith(tmpDoc.head())
-//    document.body().replaceWith(tmpDoc.body())
-//    document
+    secureDocument(document)
   }
 
   override def extractOmnitureParams(document: Document) = {
@@ -120,10 +115,6 @@ object InteractiveHtmlCleaner extends HtmlCleaner with implicits.WSRequests {
       document.head().append(html)
     }
     document
-  }
-
-  private def secureSource(src: String): String = {
-    src.replaceAllLiterally(""""//""", """"https://""").replaceAllLiterally("'//", "'https://").replaceAllLiterally("http://", "https://")
   }
 
 }
